+++ packages/ai/core/generate-text/run-tools-transformation.test.ts
@@ -16,7 +16,6 @@ it('should forward text deltas correctly', async () => {
       {
         type: 'finish',
         finishReason: 'stop',
-        logprobs: undefined,
         usage: { inputTokens: 3, outputTokens: 10 },
       },
     ]);
@@ -40,7 +39,6 @@ it('should forward text deltas correctly', async () => {
     {
       type: 'finish',
       finishReason: 'stop',
-      logprobs: undefined,
       usage: { completionTokens: 10, promptTokens: 3, totalTokens: 13 },
     },
   ]);
@@ -59,7 +57,6 @@ it('should handle immediate tool execution', async () => {
       {
         type: 'finish',
         finishReason: 'stop',
-        logprobs: undefined,
         usage: { inputTokens: 3, outputTokens: 10 },
       },
     ]);
@@ -100,7 +97,6 @@ it('should handle immediate tool execution', async () => {
     {
       type: 'finish',
       finishReason: 'stop',
-      logprobs: undefined,
       usage: { completionTokens: 10, promptTokens: 3, totalTokens: 13 },
     },
   ]);
@@ -119,7 +115,6 @@ it('should hold off on sending finish until the delayed tool result is received'
       {
         type: 'finish',
         finishReason: 'stop',
-        logprobs: undefined,
         usage: { inputTokens: 3, outputTokens: 10 },
       },
     ]);
@@ -163,7 +158,6 @@ it('should hold off on sending finish until the delayed tool result is received'
     {
       type: 'finish',
       finishReason: 'stop',
-      logprobs: undefined,
       usage: { completionTokens: 10, promptTokens: 3, totalTokens: 13 },
     },
   ]);
@@ -182,7 +176,6 @@ it('should try to repair tool call when the tool name is not found', async () =>
       {
         type: 'finish',
         finishReason: 'stop',
-        logprobs: undefined,
         usage: { inputTokens: 3, outputTokens: 10 },
       },
     ]);
@@ -234,7 +227,6 @@ it('should try to repair tool call when the tool name is not found', async () =>
     {
       type: 'finish',
       finishReason: 'stop',
-      logprobs: undefined,
       usage: { completionTokens: 10, promptTokens: 3, totalTokens: 13 },
     },
   ]);